Initial conversion of app.js to Typescript
Mostly just adding `?` and casts to get it compiling. It now works with
a couple bugs (like mute button text not updating properly), but is a
good starting point for the refactor.
There are still some implicit any and this, but trying to fix that is
basically a refactor so waiting for that to remove the tsconfig option.